requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
Pillow>=10.0
python-dateutil>=2.8
psycopg2-binary
